+++ sshd-common/src/main/java/org/apache/sshd/common/SshConstants.java
@@ -122,7 +122,8 @@ public final class SshConstants {
 
     // Some more constants
     public static final int SSH_EXTENDED_DATA_STDERR = 1;   // see RFC4254 section 5.2
-    public static final int SSH_PACKET_HEADER_LEN = 5;  // 32-bit length + 8-bit pad length
+    // 32-bit length + 8-bit pad length
+    public static final int SSH_PACKET_HEADER_LEN = Integer.BYTES + Byte.BYTES;
     /*
      * See https://tools.ietf.org/html/rfc4253#section-6.1:
      *